another change to rules
Lock down so only a new secret can be added. Existing secrets cannot be overwritten.

Process for renewal is to delete the current entry in the twofa then the cloud function will set the twfa property on the user to null.
Then next login the user will be prompted to register a new key.